+++ gear/attribute_curator/src/python/attribute_curator_app/run.py
@@ -29,10 +29,12 @@ class AttributeCuratorVisitor(GearExecutionEnvironment):
                  client: ClientWrapper,
                  project: ProjectAdaptor,
                  filename_pattern: str,
+                 curation_tag: str,
                  force_curate: bool = False):
         super().__init__(client=client)
         self.__project = project
         self.__filename_pattern = filename_pattern
+        self.__curation_tag = curation_tag
         self.__force_curate = force_curate
 
     @classmethod
@@ -56,14 +58,17 @@ class AttributeCuratorVisitor(GearExecutionEnvironment):
         proxy = client.get_proxy()
 
         filename_pattern = context.config.get('filename_pattern', "*.json")
+        curation_tag = context.config.get('curation_tag', "attribute-curator")
         force_curate = context.config.get('force_curate', False)
 
-        fw_project = get_project_from_destination(context=context, proxy=proxy)
+        #fw_project = get_project_from_destination(context=context, proxy=proxy)
+        fw_project = proxy.get_project_by_id("67f44bfc7c1a8db5864c0ad3")
         project = ProjectAdaptor(project=fw_project, proxy=proxy)
 
         return AttributeCuratorVisitor(client=client,
                                        project=project,
                                        filename_pattern=filename_pattern,
+                                       curation_tag=curation_tag,
                                        force_curate=force_curate)
 
     def run(self, context: GearToolkitContext) -> None:
@@ -82,6 +87,7 @@ class AttributeCuratorVisitor(GearExecutionEnvironment):
         run(context=context,
             deriver=deriver,
             scheduler=scheduler,
+            curation_tag=self.__curation_tag,
             force_curate=self.__force_curate)