+++ gear/regression_curator/src/python/regression_curator_app/run.py
@@ -67,7 +67,8 @@ class RegressionCuratorVisitor(GearExecutionEnvironment):
         filename_pattern = context.config.get('filename_pattern', "*UDS.json")
 
         proxy = client.get_proxy()
-        fw_project = get_project_from_destination(context=context, proxy=proxy)
+        #fw_project = get_project_from_destination(context=context, proxy=proxy)
+        fw_project = proxy.get_project_by_id("67f44bfc7c1a8db5864c0ad3")
         project = ProjectAdaptor(project=fw_project, proxy=proxy)
 
         error_outfile = context.config.get("error_outfile",